# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2.0.0
  description: Goodwill shared contract. Legacy v1 endpoints remain runnable; v2 endpoints are reserved for the parallel implementation lanes, not yet implemented.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goodwill/report:
    post:
      operationId: generateGoodwillReport
      summary: Filter the existing synthetic Upright fixture into a downloadable report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportPeriod"
      responses:
        "200":
          description: Synthetic report bytes and identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyntheticReport"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "413":
          $ref: "#/components/responses/ReportTooLarge"
...
components:
...
  responses:
...
    ReportTooLarge:
      description: Report exceeds the supervised foundation size limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ReportingError"
...
    InvalidInput:
      description: Invalid request; last successful publication is unchanged
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ReportingError"
  schemas:
...
    ReportPeriod:
      type: object
      additionalProperties: false
      required: [startDate, endDate]
      properties:
        startDate: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        endDate: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
    SyntheticReport:
      type: object
      additionalProperties: false
      required: [startDate, endDate, sourceId, reportType, filename, csv, checksum, synthetic, contractVersion]
      properties:
        startDate: { type: string }
        endDate: { type: string }
        sourceId: { type: string, enum: [upright] }
        reportType: { type: string, enum: [paid_order_items] }
        filename: { type: string, minLength: 1, maxLength: 160 }
        csv: { type: string, minLength: 1, maxLength: 250000 }
        checksum: { type: string, pattern: '^[a-f0-9]{64}$' }
        synthetic: { type: boolean, enum: [true] }
        contractVersion: { type: string, enum: ["1.0.0"] }
...
    ReportingError:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
...